third_party/blink/web_tests/platform/generic/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252_include=xhr-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest#open() assert_equals: expected "%C3%A5" but got "%E5"
Harness: the test ran to completion.
